+++ CHANGELOG.md
@@ -12,10 +12,11 @@ in 3.17. Use `PaymentSettingsInput.defaultTransactionFlowStrategy` instead.
 in 3.17. Use `PaymentSettings.defaultTransactionFlowStrategy` instead.
 
 ### GraphQL API
+- Add `storedPaymentMethodRequestDelete` mutation - #13660 by @korycins
 - Add `PaymentSettings` to `Channel` - #13677 by @korycins
 
 ### Saleor Apps
-
+- Add `STORED_PAYMENT_METHOD_DELETE_REQUESTED` webhook event - #13660 by @korycins
 - Add `NOTIFY_CUSTOMER` flag to `FulfillmentCreated` type - #13620, by @Air-t
   - Inform apps if customer should be notified when fulfillment is created.